Upgrade go dependencies
